+++ CHANGELOG.md
@@ -14,6 +14,8 @@ app.
   thanks to [@maiself](https://github.com/maiself)
   ([#2014](https://github.com/birchill/10ten-ja-reader/pull/2014)).
 - Added handling for 戶 and 內 kyūjitai.
+- Fixed unreasonable matches caused by duplicates in the deinflection reason chain
+  ([#1966](https://github.com/birchill/10ten-ja-reader/issues/1966)).
 - Fixed sorting of deinflected results in some cases (e.g. 見とれる).
 - Fixed sorting when looking up kana in some cases (e.g. なる).